+++ data/johan.txt
@@ -1,1 +1,2 @@
-E | 0 | project meeting | 2/12/2019 | 9/12/2019
+T | 0 | fly
+T | 0 | walk
