+++ src/main/java/johan/Duke.java
@@ -1,3 +1,5 @@
+package johan;
+
 public class Duke {
     public static void main(String[] args) {
         String logo = " ____        _        \n"
