+++ src/main/java/johan/command/Command.java
@@ -1,3 +1,9 @@
+package johan.command;
+
+import johan.storage.Storage;
+import johan.task.TaskList;
+import johan.ui.Ui;
+
 public abstract class Command {
     public abstract void execute(TaskList tasks, Ui ui, Storage storage) throws Exception;
     public boolean isExit() {
