+++ src/main/java/johan/command/ExitCommand.java
@@ -1,3 +1,9 @@
+package johan.command;
+
+import johan.storage.Storage;
+import johan.task.TaskList;
+import johan.ui.Ui;
+
 public class ExitCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
