+++ src/main/java/johan/command/MarkCommand.java
@@ -1,3 +1,10 @@
+package johan.command;
+
+import johan.storage.Storage;
+import johan.task.Task;
+import johan.task.TaskList;
+import johan.ui.Ui;
+
 public class MarkCommand extends Command {
     private final int taskIndex;
     private final boolean markAsDone;
