+++ src/main/java/johan/parser/Parser.java
@@ -1,3 +1,15 @@
+package johan.parser;
+
+import johan.command.Command;
+import johan.command.ExitCommand;
+import johan.command.ListCommand;
+import johan.command.MarkCommand;
+import johan.command.DeleteCommand;
+import johan.command.OnDateCommand;
+import johan.command.DeadlineCommand;
+import johan.command.EventCommand;
+import johan.command.TodoCommand;
+
 public class Parser {
     public Command parse(String input) throws Exception {
         if (input.equals("bye")) {
@@ -40,28 +52,28 @@ public class Parser {
 }
 //import java.time.LocalDate;
 //
-//public class Parser {
-//    public boolean parseAndExecute(String input, TaskList tasks, Ui ui, Storage storage) {
+//public class johan.parser.Parser {
+//    public boolean parseAndExecute(String input, johan.task.TaskList tasks, johan.ui.Ui ui, johan.storage.Storage storage) {
 //        if (input.equals("bye")) {
 //            return false;
 //        } else if (input.equals("list")) {
 //            ui.showTaskList(tasks);
 //        } else if (input.startsWith("mark ")) {
 //            int id = Integer.parseInt(input.substring(5)) - 1;
-//            Task task = tasks.getTask(id);
+//            johan.task.Task task = tasks.getTask(id);
 //            task.markAsDone();
 //            ui.showTaskMarked(task, true);
 //            storage.saveTasks(tasks.getTasks());
 //        } else if (input.startsWith("unmark ")) {
 //            int id = Integer.parseInt(input.substring(7)) - 1;
-//            Task task = tasks.getTask(id);
+//            johan.task.Task task = tasks.getTask(id);
 //            task.markAsNotDone();
 //            ui.showTaskMarked(task, false);
 //            storage.saveTasks(tasks.getTasks());
 //        } else if (input.startsWith("todo ")) {
 //            String desc = input.substring(5).trim();
 //            if (desc.isEmpty()) throw new IllegalArgumentException("The description of a todo cannot be empty.");
-//            Task task = new Todo(desc);
+//            johan.task.Task task = new johan.task.Todo(desc);
 //            tasks.addTask(task);
 //            ui.showTaskAdded(task, tasks.size());
 //            storage.saveTasks(tasks.getTasks());
@@ -70,7 +82,7 @@ public class Parser {
 //            if (byIndex == -1) throw new IllegalArgumentException("Please specify a deadline with /by.");
 //            String desc = input.substring(9, byIndex).trim();
 //            String by = input.substring(byIndex + 4).trim();
-//            Task task = new Deadline(desc, by);
+//            johan.task.Task task = new johan.task.Deadline(desc, by);
 //            tasks.addTask(task);
 //            ui.showTaskAdded(task, tasks.size());
 //            storage.saveTasks(tasks.getTasks());
@@ -81,18 +93,18 @@ public class Parser {
 //            String desc = input.substring(6, fromIndex).trim();
 //            String from = input.substring(fromIndex + 6, toIndex).trim();
 //            String to = input.substring(toIndex + 4).trim();
-//            Task task = new Event(desc, from, to);
+//            johan.task.Task task = new johan.task.Event(desc, from, to);
 //            tasks.addTask(task);
 //            ui.showTaskAdded(task, tasks.size());
 //            storage.saveTasks(tasks.getTasks());
 //        } else if (input.startsWith("delete ")) {
 //            int id = Integer.parseInt(input.substring(7)) - 1;
-//            Task task = tasks.deleteTask(id);
+//            johan.task.Task task = tasks.deleteTask(id);
 //            ui.showTaskDeleted(task, tasks.size());
 //            storage.saveTasks(tasks.getTasks());
 //        } else if (input.startsWith("on ")) {
 //            String dateStr = input.substring(3).trim();
-//            LocalDate targetDate = Johan.parseDate(dateStr);
+//            LocalDate targetDate = johan.Johan.parseDate(dateStr);
 //            ui.showTasksOnDate(tasks, targetDate);
 //        } else {
 //            throw new IllegalArgumentException("I'm sorry, but I don't know what that means :-(");
