+++ src/main/java/johan/task/Deadline.java
@@ -1,3 +1,5 @@
+package johan.task;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
