+++ src/main/java/johan/task/Event.java
@@ -1,6 +1,7 @@
+package johan.task;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
-import java.util.Date;
 
 public class Event extends Task {
