+++ src/main/java/johan/task/TaskList.java
@@ -1,3 +1,5 @@
+package johan.task;
+
 import java.util.ArrayList;
 
 public class TaskList {
