+++ src/main/java/johan/task/Todo.java
@@ -1,3 +1,5 @@
+package johan.task;
+
 public class Todo extends Task {
 
     public Todo(String description) {
